third_party/WebKit/LayoutTests/media/media-ended-expected.txt

Test ended by:

Play to the end.
When 'ended' event fires, change the source.
Verify that 'ended' event fires again on different source.
EVENT(ended)
EXPECTED (audio.playbackRate > '0') OK
EXPECTED (audio.ended == 'true') OK
RUN(audio.playbackRate = 0)
EXPECTED (audio.ended == 'true') OK
RUN(audio.playbackRate = -0.1)
EXPECTED (audio.ended == 'false') OK
EXPECTED (audio.ended == 'false') OK

EVENT(ended)
EXPECTED (audio.ended == 'true') OK
END OF TEST
